Day 26-28/requirements.txt
streamlit
extra-streamlit-components
django
djangorestframework
djangorestframework-simplejwt